refactor:remove old status field to new status
