+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/controller/ProjectParameterControllerTest.java
@@ -17,7 +17,7 @@
 
 package org.apache.dolphinscheduler.api.controller;
 
-import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.enums.v2.BaseStatus;
 import org.apache.dolphinscheduler.api.service.impl.ProjectParameterServiceImpl;
 import org.apache.dolphinscheduler.api.utils.Result;
 import org.apache.dolphinscheduler.common.enums.UserType;
@@ -50,7 +50,7 @@ public class ProjectParameterControllerTest {
         Mockito.when(projectParameterService.createProjectParameter(Mockito.any(), Mockito.anyLong(), Mockito.any(),
                 Mockito.any())).thenReturn(getSuccessResult());
         Result result = projectParameterController.createProjectParameter(loginUser, 1, "key", "value");
-        Assertions.assertEquals(Status.SUCCESS.getCode(), result.getCode());
+        Assertions.assertEquals(BaseStatus.SUCCESS.getCode(), result.getCode());
     }
 
     @Test
@@ -60,7 +60,7 @@ public class ProjectParameterControllerTest {
         Mockito.when(projectParameterService.updateProjectParameter(Mockito.any(), Mockito.anyLong(), Mockito.anyLong(),
                 Mockito.any(), Mockito.any())).thenReturn(getSuccessResult());
         Result result = projectParameterController.updateProjectParameter(loginUser, 1, 1L, "key", "value");
-        Assertions.assertEquals(Status.SUCCESS.getCode(), result.getCode());
+        Assertions.assertEquals(BaseStatus.SUCCESS.getCode(), result.getCode());
     }
 
     @Test
@@ -70,7 +70,7 @@ public class ProjectParameterControllerTest {
         Mockito.when(projectParameterService.deleteProjectParametersByCode(Mockito.any(), Mockito.anyLong(),
                 Mockito.anyLong())).thenReturn(getSuccessResult());
         Result result = projectParameterController.deleteProjectParametersByCode(loginUser, 1, 1);
-        Assertions.assertEquals(Status.SUCCESS.getCode(), result.getCode());
+        Assertions.assertEquals(BaseStatus.SUCCESS.getCode(), result.getCode());
     }
 
     @Test
@@ -80,7 +80,7 @@ public class ProjectParameterControllerTest {
         Mockito.when(projectParameterService.batchDeleteProjectParametersByCodes(Mockito.any(), Mockito.anyLong(),
                 Mockito.any())).thenReturn(getSuccessResult());
         Result result = projectParameterController.batchDeleteProjectParametersByCodes(loginUser, 1, "1");
-        Assertions.assertEquals(Status.SUCCESS.getCode(), result.getCode());
+        Assertions.assertEquals(BaseStatus.SUCCESS.getCode(), result.getCode());
     }
 
     @Test
@@ -90,7 +90,7 @@ public class ProjectParameterControllerTest {
         Mockito.when(projectParameterService.queryProjectParameterListPaging(Mockito.any(), Mockito.anyLong(),
                 Mockito.anyInt(), Mockito.anyInt(), Mockito.any())).thenReturn(getSuccessResult());
         Result result = projectParameterController.queryProjectParameterListPaging(loginUser, 1, "1", 1, 10);
-        Assertions.assertEquals(Status.SUCCESS.getCode(), result.getCode());
+        Assertions.assertEquals(BaseStatus.SUCCESS.getCode(), result.getCode());
     }
 
     @Test
@@ -100,7 +100,7 @@ public class ProjectParameterControllerTest {
         Mockito.when(projectParameterService.queryProjectParameterByCode(Mockito.any(), Mockito.anyLong(),
                 Mockito.anyLong())).thenReturn(getSuccessResult());
         Result result = projectParameterController.queryProjectParameterByCode(loginUser, 1, 1);
-        Assertions.assertEquals(Status.SUCCESS.getCode(), result.getCode());
+        Assertions.assertEquals(BaseStatus.SUCCESS.getCode(), result.getCode());
     }
 
     private User getGeneralUser() {
@@ -113,8 +113,8 @@ public class ProjectParameterControllerTest {
 
     private Result getSuccessResult() {
         Result result = new Result();
-        result.setCode(Status.SUCCESS.getCode());
-        result.setMsg(Status.SUCCESS.getMsg());
+        result.setCode(BaseStatus.SUCCESS.getCode());
+        result.setMsg(BaseStatus.SUCCESS.getMsg());
         return result;
     }
