+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/controller/TaskInstanceControllerTest.java
@@ -26,7 +26,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.enums.v2.BaseStatus;
 import org.apache.dolphinscheduler.api.service.TaskInstanceService;
 import org.apache.dolphinscheduler.api.utils.PageInfo;
 import org.apache.dolphinscheduler.api.utils.Result;
@@ -63,8 +63,8 @@ public class TaskInstanceControllerTest extends AbstractControllerTest {
         Integer pageSize = 20;
         PageInfo pageInfo = new PageInfo<TaskInstance>(pageNo, pageSize);
         result.setData(pageInfo);
-        result.setCode(Status.SUCCESS.getCode());
-        result.setMsg(Status.SUCCESS.getMsg());
+        result.setCode(BaseStatus.SUCCESS.getCode());
+        result.setMsg(BaseStatus.SUCCESS.getMsg());
 
         when(taskInstanceService.queryTaskListPaging(any(), eq(1L), eq(1), eq(""), eq(""), eq(""), any(), eq(""), any(),
                 any(),
@@ -73,7 +73,7 @@ public class TaskInstanceControllerTest extends AbstractControllerTest {
         Result taskResult = taskInstanceController.queryTaskListPaging(null, 1L, 1, "", "", "",
                 "", 1L, "", TaskExecutionStatus.SUCCESS, "192.168.xx.xx", "2020-01-01 00:00:00", "2020-01-02 00:00:00",
                 TaskExecuteType.BATCH, pageNo, pageSize);
-        Assertions.assertEquals(Integer.valueOf(Status.SUCCESS.getCode()), taskResult.getCode());
+        Assertions.assertEquals(Integer.valueOf(BaseStatus.SUCCESS.getCode()), taskResult.getCode());
     }
 
     @Disabled
@@ -83,7 +83,7 @@ public class TaskInstanceControllerTest extends AbstractControllerTest {
         paramsMap.add("taskInstanceId", "104");
 
         Result mockResult = new Result();
-        putMsg(mockResult, Status.SUCCESS);
+        putMsg(mockResult, BaseStatus.SUCCESS);
         when(taskInstanceService.forceTaskSuccess(any(User.class), anyLong(), anyInt())).thenReturn(mockResult);
 
         MvcResult mvcResult = mockMvc.perform(post("/projects/{projectName}/task-instance/force-success", "cxc_1113")
@@ -94,7 +94,7 @@ public class TaskInstanceControllerTest extends AbstractControllerTest {
                 .andReturn();
 
         Result result = JSONUtils.parseObject(mvcResult.getResponse().getContentAsString(), Result.class);
-        Assertions.assertEquals(Status.SUCCESS.getCode(), result.getCode().intValue());
+        Assertions.assertEquals(BaseStatus.SUCCESS.getCode(), result.getCode().intValue());
     }
 
 }
