+++ dolphinscheduler-api/src/test/java/org/apache/dolphinscheduler/api/controller/UiPluginControllerTest.java
@@ -25,7 +25,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import org.apache.dolphinscheduler.api.enums.Status;
+import org.apache.dolphinscheduler.api.enums.v2.BaseStatus;
 import org.apache.dolphinscheduler.api.service.UiPluginService;
 import org.apache.dolphinscheduler.api.utils.Result;
 import org.apache.dolphinscheduler.common.constants.Constants;
@@ -51,7 +51,7 @@ public class UiPluginControllerTest extends AbstractControllerTest {
     private static final Result expectResponseContent = JSONUtils.parseObject(
             "{\"code\":0,\"msg\":\"success\",\"data\":\"Test Data\",\"success\":true,\"failed\":false}", Result.class);
     private static final ImmutableMap<String, Object> uiPluginServiceResult =
-            ImmutableMap.of(Constants.STATUS, Status.SUCCESS, Constants.DATA_LIST, "Test Data");
+            ImmutableMap.of(Constants.STATUS, BaseStatus.SUCCESS, Constants.DATA_LIST, "Test Data");
 
     @MockBean(name = "uiPluginService")
     private UiPluginService uiPluginService;
